Update get_signal3g.py
Please update as follows: We need to have a float, so we can use the comma too, not integer for the signal, Please check the values we ask in the command:
signal_str_val = result_string[16:20]
Maybe we have different comgt? With your parameters [15:17] I just can not get anything.
Thank you!